// src/lib/types/league.d.ts
export type LeagueData = {
  idLeague: string | null;
  strLeagueAlternate: string | null;
  strCountry: string | null;
  strCurrentSeason: string | null;
  strDescriptionEN: string | null;
  strBadge: string | null;
  strBanner: string | null;
  idSoccerXML: string | null;
  idAPIfootball: string | null;
  strSport: string | null;
  intDivision: string | null;
  idCup: string | null;
  intFormedYear: string | null;
  dateFirstEvent: string | null;
  strGender: string | null;
  strWebsite: string | null;
  strFacebook: string | null;
  strInstagram: string | null;
  strTwitter: string | null;
  strYoutube: string | null;
  strRSS: string | null;
  strDescriptionDE: string | null;
  strDescriptionFR: string | null;
  strDescriptionIT: string | null;
  strDescriptionCN: string | null;
  strDescriptionJP: string | null;
  strDescriptionRU: string | null;
  strDescriptionES: string | null;
  strDescriptionPT: string | null;
  strDescriptionSE: string | null;
  strDescriptionNL: string | null;
  strDescriptionHU: string | null;
  strDescriptionNO: string | null;
  strDescriptionPL: string | null;
  strDescriptionIL: string | null;
  strTvRights: string | null;
  strFanart1: string | null;
  strFanart2: string | null;
  strFanart3: string | null;
  strFanart4: string | null;
  strLogo: string | null;
  strPoster: string | null;
  strTrophy: string | null;
  strNaming: string | null;
  strComplete: string | null;
  strLocked: string | null;
};